processor/requirements.txt
kafka-python==2.0.2
python-dotenv==0.19.2
snowflake-connector-python==2.7.12
nltk==3.6.7
textblob==0.17.1
pandas==1.5.0